+++ package.json
@@ -27,6 +27,7 @@
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18",
     "react-feather": "^2.0.10",
+    "sharp": "^0.33.4",
     "uuid": "^10.0.0",
     "zustand": "^4.5.2"
   },
