+++ src/app/chef/page.tsx
@@ -3,7 +3,6 @@ import React, { useEffect } from 'react'
 import { orderListRouter } from '@/lib/database/orderListRouter'
 import { OrderStore } from '@/lib/store/orderStore'
 import ChefOrderCard from '../../ui/component/ProductList/ChefOrderCard'
-import Header from '@/ui/component/NavigationBar/Header'
 
 export default function ChefPage() {
   const { orderList } = OrderStore()
@@ -13,8 +12,6 @@ export default function ChefPage() {
   }, [getOrderList])
 
   return (
-    <>
-      <Header />
       <div className='mini-h-[80svh] grid grid-cols-3 gap-y-3 pt-28'>
         {
           orderList.map((list, index) => (
@@ -26,6 +23,5 @@ export default function ChefPage() {
           ))
         }
       </div>
-    </>
   )
 }
