+++ src/app/owner/layout.tsx
@@ -1,14 +1,16 @@
-"use client"
 import Sidebar from '@/ui/component/NavigationBar/Sidebar';
+import { Metadata } from 'next';
 import React from 'react';
 
-const AboutLayout = ({ children }: { children: React.ReactNode }) => {
+export const metadata:Metadata = {
+    title: "Owner"
+}
+
+export default function AboutLayout({ children }: { children: React.ReactNode }) {
     return (
-        <main>
+        <>
             <Sidebar />
             <section className='pt-24'>{children}</section>
-        </main>
+        </>
     );
 };
-
-export default AboutLayout;
