+++ src/app/page.tsx
@@ -4,7 +4,7 @@ import { Alert, AlertIcon, Button } from "@chakra-ui/react";
 import { useEffect } from "react";
 import PreviewAppList from "../ui/component/PreviewApps/PreviewAppList";
 import Header from "../ui/component/NavigationBar/HomeHeader";
-import FadeInUp from "../ui/framer-motion/FadeInUp";
+import FadeInUp from "../ui/framer-motion/Animation/FadeInUp";
 
 export default function Home() {
   useEffect(() => {
