+++ src/ui/component/Logo/Logo.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import { Ubuntu } from "next/font/google"
+import Link from 'next/link'
 
 const ubuntu = Ubuntu({ subsets: ["latin"], weight: ["500", "700"] })
 
@@ -8,5 +9,9 @@ type Props = {
 }
 
 export default function Logo({ className }: Props) {
-    return <h1 className={`${ubuntu.className} text-2xl ${className}`}>Warung<span className='text-primary-blue'>Web</span></h1>
+    return (
+        <Link href={"/"}>
+            <h1 className={`${ubuntu.className} text-2xl ${className}`}>Warung<span className='text-primary-blue'>Web</span></h1>
+        </Link>
+    )
 }
