+++ src/ui/component/NavigationBar/CashierNavbar.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
-import SearchBar from '../../elements/SearchBar'
 import GroceryListButton from '../GroceryList/GroceryListButton'
 import ProductCategoryTabs from '../ProductCategoryTab/ProductCategoryTabs'
 import Logo from '../Logo/Logo'
+import Clicked from '@/ui/framer-motion/Animation/Clicked'
+import SearchProduct from '../SearchBar/SearchProduct'
 
 export default function Navbar() {
     return (
@@ -12,8 +13,10 @@ export default function Navbar() {
                     <Logo />
                 </div>
                 <div className='w-2/3 sm:w-1/2 lg:w-1/3 flex gap-2 sm:gap-3'>
-                    <SearchBar />
-                    <GroceryListButton />
+                    <SearchProduct />
+                    <Clicked scale={0.95}>
+                        <GroceryListButton />
+                    </Clicked>
                 </div>
             </section>
             <section className='w-fit sm:w-full h-1/2 pl-7 flex items-end sm:items-center'>
