+++ src/ui/component/NavigationBar/Header.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
-import LogoBlack from '../Logo/LogoBlack'
+import Logo from '../Logo/Logo'
 
 export default function Header() {
   return (
     <header className='h-32 w-full flex justify-center items-center'>
-      <LogoBlack/>
+      <Logo/>
     </header>
   )
 }
